# scripts/openapi/openapi_latest.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenScan3 API v0.9",
    "version": "0.9"
  },
  "paths": {
...
    "/tasks/": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Get All Tasks",
        "description": "Retrieve a list of all tasks known to the task manager.\n\nReturns:\n    List[Task]: A list of all tasks known to the task manager.",
        "operationId": "get_all_tasks",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Task"
                  },
                  "type": "array",
                  "title": "Response Get All Tasks Tasks  Get"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/tasks/{task_id}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Get Task Status",
        "description": "Retrieve the status and details of a specific task.\n\nArgs:\n    task_id: The ID of the task to retrieve.\n\nReturns:\n    Task: The task object with its status and details.",
        "operationId": "get_task_status",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "tasks"
        ],
        "summary": "Cancel Task",
        "description": "Request cancellation of a running task.\n\nArgs:\n    task_id: The ID of the task to cancel.\n\nReturns:\n    Task: The task object with its status and details.",
        "operationId": "cancel_task",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}/cleanup": {
      "delete": {
        "tags": [
          "tasks"
        ],
        "summary": "Delete a terminal task record",
        "description": "Remove a terminal task from persistence and memory.",
        "operationId": "delete_task",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}/pause": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Pause a Task",
        "description": "Pauses a running task.\n\nArgs:\n    task_id: The ID of the task to pause.\n\nReturns:\n    Task: The task object with its status and details.",
        "operationId": "pause_task",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}/resume": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Resume a Task",
        "description": "Resumes a paused or interrupted task.\n\nArgs:\n    task_id: The ID of the task to resume.\n\nReturns:\n    Task: The task object with its status and details.",
        "operationId": "resume_task",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_name}": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Create Task",
        "description": "Create and start a new background task with optional parameters.\n\nThe request body accepts:\n- **depends_on**: Optional ID of a prerequisite task\n- **args**: List of positional arguments (e.g., `[\"project_name\", 0]`)\n- **kwargs**: Dictionary of keyword arguments (e.g., `{\"num_batches\": 5}`)\n\nArgs:\n    task_name: The name of the task to create, as registered in the TaskManager.\n    args: Positional arguments to pass to the task's run method.\n    kwargs: Keyword arguments to pass to the task's run method.\n\nReturns:\n    The created task object.\n\nExamples:\n    ```json\n    // No parameters\n    {}\n\n    // With positional args\n    {\n        \"args\": [\"MyProject\", 0]\n    }\n\n    // With keyword args\n    {\n        \"kwargs\": {\"num_calibration_batches\": 5}\n    }\n\n    // With both\n    {\n        \"args\": [\"MyProject\", 0],\n        \"kwargs\": {\"num_calibration_batches\": 5}\n    }\n    ```",
        "operationId": "create_task",
        "parameters": [
          {
            "name": "task_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Name"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_task_tasks__task_name__post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_create_task_tasks__task_name__post": {
        "properties": {
          "args": {
            "items": {},
            "type": "array",
            "title": "Args",
            "description": "Positional arguments for the task",
            "default": []
          },
          "kwargs": {
            "additionalProperties": true,
            "type": "object",
            "title": "Kwargs",
            "description": "Keyword arguments for the task",
            "default": {}
          },
          "depends_on": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Depends On",
            "description": "Optional task ID that must complete successfully before this task runs"
          }
        },
        "type": "object",
        "title": "Body_create_task_tasks__task_name__post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Task": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "task_type": {
            "type": "string",
            "title": "Task Type"
          },
          "is_exclusive": {
            "type": "boolean",
            "title": "Is Exclusive",
            "description": "Whether this task is exclusive and should not run concurrently",
            "default": false
          },
          "is_blocking": {
            "type": "boolean",
            "title": "Is Blocking",
            "description": "Whether this task is blocking and should run in a separate thread",
            "default": false
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus",
            "default": "pending"
          },
          "progress": {
            "$ref": "#/components/schemas/TaskProgress"
          },
          "result": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "depends_on": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Depends On",
            "description": "Task ID that must complete successfully before this task may run."
          },
          "run_args": {
            "items": {},
            "type": "array",
            "title": "Run Args",
            "description": "Positional arguments the task was started with."
          },
          "run_kwargs": {
            "additionalProperties": true,
            "type": "object",
            "title": "Run Kwargs",
            "description": "Keyword arguments the task was started with."
          }
        },
        "type": "object",
        "required": [
          "name",
          "task_type"
        ],
        "title": "Task",
        "description": "Represents a background task."
      },
      "TaskProgress": {
        "properties": {
          "current": {
            "type": "number",
            "title": "Current",
            "description": "The current step or value of progress (e.g., files processed).",
            "default": 0.0
          },
          "total": {
            "type": "number",
            "title": "Total",
            "description": "The total number of steps or value for completion (e.g., total files).",
            "default": 0.0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          }
        },
        "type": "object",
        "title": "TaskProgress",
        "description": "Model for task progress."
      },
      "TaskStatus": {
        "type": "string",
        "enum": [
          "pending",
          "running",
          "paused",
          "completed",
          "cancelled",
          "error",
          "interrupted"
        ],
        "title": "TaskStatus",
        "description": "Enum for task status"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}